Visual Studios 2013 MSI Bug
I created a script that detects the MSI files and alerts you and gathers the logs, counts them and collects size data and stores them in a log file. The other script attempts to fix it (Requires Elevated Permission) by cleaning up the log files and attempt to fix the issue by creating the URTInstallPath_GAC folder in NetFramework directory.